news

+++ src/Controller/NouveauteController.php
@@ -167,17 +167,28 @@ class NouveauteController extends AbstractController
         $user=$this->getUser();
         $nouveaute= $this->getDoctrine()->getRepository('App:Nouveaute')->find($id);
 
-        if($user->getNomManager()==$nouveaute->getAuteurNom() and $user->getPrenomManager()==$nouveaute->getAuteurPreom()  and $nouveaute->getTypenouveaute()==1)
+        if($user instanceof Manager)
         {
-            $oldImg = $nouveaute->getLienImage();
-            @unlink($fileUploader->getTargetDirectory . $oldImg);
+            if($user->getNomManager()==$nouveaute->getAuteurNom() and $user->getPrenomManager()==$nouveaute->getAuteurPreom()  and $nouveaute->getTypenouveaute()==1)
+            {
+                $oldImg = $nouveaute->getLienImage();
+                @unlink($fileUploader->getTargetDirectory . $oldImg);
 
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($nouveaute);
-            $entityManager->flush();
-            return $this->redirectToRoute('nouveaute');
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($nouveaute);
+                $entityManager->flush();
+                return $this->redirectToRoute('nouveaute');
+            }
         }
 
+
+        $oldImg = $nouveaute->getLienImage();
+        @unlink($fileUploader->getTargetDirectory . $oldImg);
+
+        $entityManager = $this->getDoctrine()->getManager();
+        $entityManager->remove($nouveaute);
+        $entityManager->flush();
+        return $this->redirectToRoute('nouveaute');
 
     }
 }
